cloud-run/ml-api/app/requirements.txt
Flask
torch
transformers
sentencepiece
gunicorn
google-auth
google.cloud
